+++ lib/go-tc/enum.go
@@ -176,6 +176,9 @@ const DSProtocolHTTPS = 1
 const DSProtocolHTTPAndHTTPS = 2
 const DSProtocolHTTPToHTTPS = 3
 
+// CacheInterfacesAggregate represent the interface that stores aggregate data for each interface on the cache
+const CacheInterfacesAggregate = "aggregate"
+
 // CacheStatus represents the Traffic Server status set in Traffic Ops (online, offline, admin_down, reported). The string values of this type should match the Traffic Ops values.
 type CacheStatus string