+++ traffic_ops/traffic_ops_golang/topology/validation.go
@@ -53,17 +53,39 @@ func checkForSelfParents(nodes []tc.TopologyNode, index int) error {
 	return nil
 }
 
-func checkForEdgeParents(nodes []tc.TopologyNode, cachegroups []tc.CacheGroupNullable, nodeIndex int) error {
-	node := nodes[nodeIndex]
+// checkForEdgeParents returns an error if an index given in the parents array, adds a warning + returns a nil error if
+// an edge parents an edge, and returns an error if an edge parents a non-edge cachegroup.
+func (topology *TOTopology) checkForEdgeParents(cacheGroups []tc.CacheGroupNullable, nodeIndex int) error {
+	node := topology.Nodes[nodeIndex]
 	errs := make([]error, len(node.Parents))
-	for parentIndex, cachegroupIndex := range node.Parents {
-		if cachegroupIndex < 0 || cachegroupIndex >= len(cachegroups) {
-			errs = append(errs, fmt.Errorf("parent %d of cachegroup %s refers to a cachegroup at index %d, but no such cachegroup exists", parentIndex, node.Cachegroup, cachegroupIndex))
+	for parentIndex, parentCacheGroupIndex := range node.Parents {
+		if parentCacheGroupIndex < 0 || parentCacheGroupIndex >= len(topology.Nodes) {
+			errs = append(errs, fmt.Errorf("parent %d of cachegroup %s refers to a cachegroup at index %d, but no such cachegroup exists", parentIndex, node.Cachegroup, parentCacheGroupIndex))
 			break
 		}
-		cacheGroupType := cachegroups[cachegroupIndex].Type
-		if *cacheGroupType == tc.CacheGroupEdgeTypeName {
-			errs = append(errs, fmt.Errorf("cachegroup %v's type is %v; it cannot be a parent of %v", nodes[cachegroupIndex].Cachegroup, tc.CacheGroupEdgeTypeName, node.Cachegroup))
+		parentCacheGroupType := *cacheGroups[parentCacheGroupIndex].Type
+		if parentCacheGroupType != tc.CacheGroupEdgeTypeName {
+			continue
+		}
+		switch cacheGroupType := *cacheGroups[nodeIndex].Type; cacheGroupType {
+		case tc.CacheGroupEdgeTypeName:
+			parentTerm := "parent"
+			if parentIndex == 1 {
+				parentTerm = "secondary " + parentTerm
+			}
+			topology.Alerts.AddNewAlert(tc.WarnLevel, fmt.Sprintf(
+				"%s-typed cachegroup %s is a %s of %s, unexpected behavior may result",
+				parentCacheGroupType,
+				topology.Nodes[parentCacheGroupIndex].Cachegroup,
+				parentTerm,
+				node.Cachegroup))
+		default:
+			errs = append(errs, fmt.Errorf(
+				"cachegroup %s's type is %s; it cannot parent a %s-typed cachegroup %s",
+				topology.Nodes[parentCacheGroupIndex].Cachegroup,
+				parentCacheGroupType,
+				cacheGroupType,
+				node.Cachegroup))
 		}
 	}
 	return util.JoinErrs(errs)